@@ -4,27 +4,14 @@
 //
 //  Created by Raymond Pinto on 20.12.19.
 //
-
 import SwiftUI
 
 
 // MARK: FeedbackView
 /// This View holds the markup image and the feedback text field for the user to send feedback.
 struct FeedbackView: View {
-    // MARK: ActiveSheet
-    /// The type of active sheet that can be presented by the `FeedbackView`
-    enum ActiveSheet {
-        /// The login sheet is presented
-        case loginSheet
-        /// The markup sheet is presented
-        case markupSheet
-    }
-    
-    
     /// The instance of the Observable Object class named Model,  to share state data anywhere it’s needed.
     @EnvironmentObject var state: PrototyperState
-    /// <#Description#>
-    @EnvironmentObject var apiHandler: APIHandler
     
     /// This State variable holds the feedback text.
     @State var descriptionText: String = ""
@@ -32,8 +19,6 @@ struct FeedbackView: View {
     @State var showSheet: Bool = false
     /// This State variable is used to tell the View if the screenshot needs to be sent as feedback or not.
     @State var showScreenshot: Bool = true
-    /// The active Sheet decides if the login View or the SendFeedback View needs to be shown when the showSheet variable is updated.
-    @State var activeSheet: ActiveSheet = .loginSheet
     /// This State variable when set to true sends the Feedback to Prototyper.
     @State var showSendFeedbackView: Bool = false
     /// The variable holds the image and the feedback text to be sent.
@@ -59,19 +44,10 @@ struct FeedbackView: View {
             .navigationBarTitle("Give Feedback")
             .navigationBarItems(leading: cancelButton, trailing: sendButton)
             .sheet(isPresented: $showSheet) {
-                if self.activeSheet == .loginSheet {
-                    NavigationView {
-                        LoginView(finishLoggingIn: self.$showSendFeedbackView)
-                    }
-                        .environmentObject(self.state)
-                        .environmentObject(self.apiHandler)
-                } else {
                     NavigationView {
                         EditScreenshotView()
                     }
                         .environmentObject(self.state)
-                        .environmentObject(self.apiHandler)
-                }
             }
         }
     }
@@ -80,7 +56,7 @@ struct FeedbackView: View {
     var screenshot: some View {
         ZStack(alignment: .topTrailing) {
             ZStack {
-                Image(uiImage: state.screenshotWithMarkup!)
+                Image(uiImage: state.getScreenshotWithMarkup())
                     .resizable()
                     .shadow(color: Color.primary.opacity(0.2), radius: 5.0)
                     .scaledToFit()
@@ -103,7 +79,7 @@ struct FeedbackView: View {
     /// Holds the screenshot and the feedback text to be sent to Prototyper
     var currentFeedback: Feedback {
         var creatorName: String?
-        if !apiHandler.isLoggedIn {
+        if state.apiHandler.userIsLoggedIn {
             creatorName = UserDefaults.standard.string(forKey: UserDefaultKeys.username)
         }
         
@@ -130,25 +106,18 @@ struct FeedbackView: View {
     /// Dismisses the view and makes the Feedback bubble appear again.
     private func cancel() {
         Prototyper.dismissView()
-        Prototyper.currentState.feedbackButtonIsHidden = !Prototyper.settings.showFeedbackButton
+        state.setFeedbackButtonIsHidden()
     }
     
     /// Sends the feedback to Prototyper via the SendFeedbackView
     private func send() {
         feedback = currentFeedback
-        if apiHandler.isLoggedIn || state.continueWithoutLogin {
-            state.continueWithoutLogin = false
-            self.showSendFeedbackView = true
-        } else {
-            activeSheet = .loginSheet
-            self.showSheet = true
-        }
+        self.showSendFeedbackView = true
     }
     
     /// Brings up the EditScreenShotView to Markup the screenshot
     private func editImage() {
-        activeSheet = .markupSheet
-        showSheet = true
+            showSheet = true
     }
     
     /// Deletes the screenshot from the View
