@@ -23,19 +23,24 @@ struct LoginView: View {
             static let withLoginText = "Continue with login"
         }
         
+        enum DescrpitionText {
+            static let withoutLoginText = "Enter a name"
+            static let withLoginText = "Please log in using your Prototyper/TUM credentials to send feedback"
+        }
+        
+        enum ContinueButtonText {
+            static let withoutLoginText = "Continue"
+            static let withLoginText = "Login"
+        }
+        
         static let alertText = "Could not log in! Please check your login credentials and try again."
     }
     
     
     /// The instance of the Observable Object class named Model,  to share state data anywhere it’s needed.
     @EnvironmentObject var state: PrototyperState
-    /// <#Description#>
-    @EnvironmentObject var apiHandler: APIHandler
     /// Environment variable for presentationMode to dismiss the View.
     @Environment(\.presentationMode) var presentationMode
-    
-    /// A binding variable to update the calling view once the user has clicked the login button.
-    @Binding var finishLoggingIn: Bool
     
     /// The username of the user.
     @State var userid: String = ""
@@ -45,57 +50,81 @@ struct LoginView: View {
     @State var showLoginErrorAlert: Bool = false
     /// A boolean to check if the user is proceeding without logging in,
     @State var continueWithoutLogin: Bool = false
+    ///Shows ProgessView while logging in
+    @State var isLoginIn: Bool = false
+    
+    var descriptionText: String {
+        if continueWithoutLogin {
+            return LoginViewConstants.DescrpitionText.withoutLoginText
+        } else {
+            return LoginViewConstants.DescrpitionText.withLoginText
+        }
+    }
+    
+    var continueButtonText: String {
+        if continueWithoutLogin {
+            return LoginViewConstants.ContinueButtonText.withoutLoginText
+        } else {
+            return LoginViewConstants.ContinueButtonText.withLoginText
+        }
+    }
     
     
     /// The text to be displayed as placeholder text on the text field.
     var userNamePlaceHolder: String {
         if continueWithoutLogin {
-              return LoginViewConstants.UserNamePlaceHolder.withLoginText
+            return LoginViewConstants.UserNamePlaceHolder.withLoginText
         } else {
-              return LoginViewConstants.UserNamePlaceHolder.withoutLoginText
+            return LoginViewConstants.UserNamePlaceHolder.withoutLoginText
         }
     }
     
     /// The text to be displayed on the Login button.
     var subButtonText: String {
         if continueWithoutLogin {
-              return LoginViewConstants.SubButtonText.withLoginText
+            return LoginViewConstants.SubButtonText.withoutLoginText
         } else {
-              return LoginViewConstants.SubButtonText.withoutLoginText
+            return LoginViewConstants.SubButtonText.withLoginText
         }
     }
     
     var body: some View {
-            VStack {
-                VStack(alignment: .leading) {
-                    Text("Please log in using your Prototyper/TUM credentials to send feedback")
-                        .lineLimit(2)
-                }.padding()
-                    
-                VStack(spacing: 25) {
-                    TextField(userNamePlaceHolder, text: $userid)
-                    if !continueWithoutLogin {
-                        SecureField("Password", text: $password)
+        VStack {
+            VStack(alignment: .leading) {
+                Text(descriptionText)
+                    .lineLimit(2)
+            }.padding()
+            
+            VStack(spacing: 25) {
+                TextField(userNamePlaceHolder, text: $userid)
+                if !continueWithoutLogin {
+                    SecureField("Password", text: $password)
+                }
+                Button(action: login) {
+                    VStack {
+                        if isLoginIn {
+                            ProgressView()
+                        } else {
+                            Text(continueButtonText).bold()
+                        }
                     }
-                    Button(action: login) {
-                        Text("Login").bold()
-                        .foregroundColor(.white)
-                        .padding(8)
-                            .frame(maxWidth: .infinity, alignment: .center)
-                            .background(buttonColor)
-                        .cornerRadius(10)
-                    }.disabled(userid.isEmpty || password.isEmpty)
-                    Button(action: loginSubButton) {
-                        Text(subButtonText).bold()
-                    }.alert(isPresented: $showLoginErrorAlert) {
-                        Alert(title: Text("Error"), message: Text(LoginViewConstants.alertText), dismissButton: .default(Text("OK")))
-                    }
-                    Spacer()
-                }.padding()
-            }
-            .navigationBarTitle("Sign In")
-            .navigationBarBackButtonHidden(true)
-            .navigationBarItems(leading: backButton)
+                    .foregroundColor(.white)
+                    .padding(8)
+                    .frame(maxWidth: .infinity, alignment: .center)
+                    .background(buttonColor)
+                    .cornerRadius(10)
+                }.disabled(userid.isEmpty || password.isEmpty)
+                Button(action: loginSubButton) {
+                    Text(subButtonText).bold()
+                }.alert(isPresented: $showLoginErrorAlert) {
+                    Alert(title: Text("Error"), message: Text(LoginViewConstants.alertText), dismissButton: .default(Text("OK")))
+                }
+                Spacer()
+            }.padding()
+        }
+        .navigationBarTitle("Sign In")
+        .navigationBarBackButtonHidden(true)
+        .navigationBarItems(leading: backButton)
     }
     
     /// The color switch to show that the button is enabled.
@@ -121,16 +150,19 @@ struct LoginView: View {
     private func login() {
         if continueWithoutLogin {
             UserDefaults.standard.set(userid, forKey: UserDefaultKeys.username)
-            state.continueWithoutLogin = true
-            self.finishLoggingIn = true
-            self.presentationMode.wrappedValue.dismiss()
+            state.apiHandler.continueWithoutLogin = true
         } else {
-            apiHandler.login(userid,
-                             password: password,
-                             success: { self.presentationMode.wrappedValue.dismiss()
-            }, failure: { _ in
-                self.showingLoginErrorAlert()
-            })
+            isLoginIn = true
+            state.apiHandler.login(userid,
+                                   password: password,
+                                   success: {
+                                    isLoginIn = false
+                                   },
+                                   failure: { _ in
+                                    isLoginIn = false
+                                    self.showingLoginErrorAlert()
+                                   })
+            state.apiHandler.continueWithoutLogin = false
         }
     }
     
