@@ -1,33 +1,32 @@
 //
 //  Prototyper.swift
-//  
+//
 //
 //  Created by Paul Schmiedmayer on 4/24/20.
 //
 
 import UIKit
+import PencilKit
 
-
-/// <#Description#>
+/// Prototyper settings is a struct that contains information about the configuration of the prototyper framework and is used for the initialization
 public struct PrototyperSettings {
-    /// <#Description#>
+    /// The default setting establishes a connection to https://prototyper.ase.in.tum.de/login and shows the feedback button
     public static var `default`: PrototyperSettings = {
         PrototyperSettings(showFeedbackButton: true,
                            prototyperInstance: URL(string: "https://prototyper.ase.in.tum.de/login")!)
-        // swiftlint:disable:previous force_unwrap
     }()
     
     
-    /// <#Description#>
+    /// Boolean whether the feedback button is shown
     public var showFeedbackButton: Bool
-    /// <#Description#>
+    /// URL for the prototyper instance
     public var prototyperInstance: URL
     
     
-    /// <#Description#>
+    /// The initializer to creat a new `PrototyperSettings` struct
     /// - Parameters:
-    ///   - showFeedbackButton: <#showFeedbackButton description#>
-    ///   - prototyperInstance: <#prototyperInstance description#>
+    ///   - showFeedbackButton: Boolean if the feedback button is shown
+    ///   - prototyperInstance: URL for the prototyper instance
     public init(showFeedbackButton: Bool = `default`.showFeedbackButton,
                 prototyperInstance: URL = `default`.prototyperInstance) {
         self.showFeedbackButton = showFeedbackButton
@@ -36,9 +35,9 @@ public struct PrototyperSettings {
 }
 
 
-/// <#Description#>
-public class PrototyperState: ObservableObject {
-    /// <#Description#>
+/// This class contains the information which are shared a cross the different classes
+class PrototyperState: ObservableObject {
+    /// Boolean whether the feedback button is shown
     @Published var feedbackButtonIsHidden: Bool
     /// The screenshot that should be displayed in the feedback view
     @Published var screenshot: UIImage? {
@@ -49,29 +48,53 @@ public class PrototyperState: ObservableObject {
     /// The screenshot that should be displayed in the feedback view including the rendered markup
     @Published var screenshotWithMarkup: UIImage?
     /// This variable holds all the drawings the user draws in the Markup view.
-    @Published var markupDrawings: [Drawing] = [Drawing]()
-    /// This boolean variable is used to check if the user is submitting feedback with or without logging in.
-    @Published var continueWithoutLogin: Bool = false
+    @Published var markupDrawings: PKDrawing?
+    ///The APIHandler handels everything related to networking
+    @Published var apiHandler: APIHandler!
     
+    /// Initializer for the state
+    /// - Parameter feedbackButtonIsHidden: Boolean whether the feedback button is shown
+    init(_ settings: PrototyperSettings) {
+        self.feedbackButtonIsHidden = !settings.showFeedbackButton
+        apiHandler = APIHandler(prototyperInstance: settings.prototyperInstance, state: self)
+    }
     
-    /// <#Description#>
-    /// - Parameter feedbackButtonIsHidden: <#feedbackButtonIsHidden description#>
-    init(feedbackButtonIsHidden: Bool) {
+    ///Function to set FeedbackButtonIsHidden
+    func setFeedbackButtonIsHidden(_ feedbackButtonIsHidden: Bool) {
         self.feedbackButtonIsHidden = feedbackButtonIsHidden
+    }
+    
+    ///Function to set FeedbackButtonIsHidden depending on the PrototyperState
+    func setFeedbackButtonIsHidden() {
+        self.feedbackButtonIsHidden = !Prototyper.settings.showFeedbackButton
+    }
+    
+    ///Returns the Screenshot if available
+    func getScreenshot() -> UIImage {
+        guard let screenshot = screenshot else {
+            perror("Could not load the screenshot!")
+            return UIImage()
+        }
+        return screenshot
+    }
+    
+    ///Returns the Screenshot if available
+    func getScreenshotWithMarkup() -> UIImage {
+        guard let screenshotWithMarkup = screenshotWithMarkup else {
+            perror("Could not load the screenshot!")
+            return UIImage()
+        }
+        return screenshotWithMarkup
     }
 }
 
 
-/// <#Description#>
+/// Main class of the framework
 public class Prototyper {
-    /// <#Description#>
+    /// Prototyper settings is a struct that contains information about the configuration of the prototyper framework and is used for the initialization
     static var settings: PrototyperSettings!
-    /// <#Description#>
+    /// This class contains the information which are shared a cross the different classes
     static var currentState: PrototyperState!
-    /// <#Description#>
-    static var apiHandler: APIHandler!
-    
-    
     /// The topmost View where the bubble was pressed.
     static var topViewController: UIViewController? {
         guard let rootViewController = UIApplication.shared.windows.first?.rootViewController else {
@@ -105,23 +128,17 @@ public class Prototyper {
     }
     
     
-    /// <#Description#>
-    /// - Parameter settings: <#settings description#>
+    /// This static function is used to initialize the prototyper framework with a suitable `PrototyperSettings` object
+    /// - Parameter settings: Struct that contains information about the configuration of the prototyper framework
     public static func configure(_ settings: PrototyperSettings) {
+        self.currentState = PrototyperState(settings)
         self.settings = settings
-        
-        self.currentState = PrototyperState(feedbackButtonIsHidden: !settings.showFeedbackButton)
-        
-        self.apiHandler = APIHandler(settings: settings)
-        apiHandler.tryToFetchReleaseInfos()
-        apiHandler.tryToLogin()
-        
         if settings.showFeedbackButton {
             addFeedbackBubble()
         }
     }
     
-    
+    @available(*, unavailable)
     private init() {
         fatalError("An instance of Prototyper should never be created")
     }
